% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesintro.R
\docType{package}
\name{bayesintro}
\alias{bayesintro}
\title{bayesintro: Introduction to Bayesian Statistics}
\description{
The \code{bayesintro} package provides a collection of
interactive \code{learnr} materials for the workshop
"Introduction to Bayesian Statistics". You can call the
materials for lectures and labs using the code provided in the examples
below.
}
\examples{
\dontrun{
  # Introduction
  learnr::run_tutorial("00-00-int", package = "bayesintro")

  # Day 1
  learnr::run_tutorial("01-01-lec", package = "bayesintro")
  learnr::run_tutorial("01-02-lec", package = "bayesintro")
  learnr::run_tutorial("01-03-lab", package = "bayesintro")

  # Day 2
  learnr::run_tutorial("02-01-lec", package = "bayesintro")
  learnr::run_tutorial("02-02-lec", package = "bayesintro")
  learnr::run_tutorial("02-03-lab", package = "bayesintro")

  # Day 3
  learnr::run_tutorial("03-01-lab", package = "bayesintro")
}
}
\seealso{
\code{\link{learnr}}
}
